重构框架
