+++ DotNETBuild/Context/AppConfigurator.cs
@@ -19,12 +19,26 @@ namespace dotnetCampus.DotNETBuild.Context
         {
             if (_appConfigurator != null)
             {
-                throw new Exception("必须在调用 GetAppConfigurator 方法之前设置配置文件路径，请将设置路径的代码放在程序运行最前");
+                throw new InvalidOperationException("必须在调用 GetAppConfigurator 方法之前设置配置文件路径，请将设置路径的代码放在程序运行最前");
             }
 
             _appConfigurator = ConfigurationFactory.FromFile(file.FullName).CreateAppConfigurator();
         }
 
+        internal static void SetAppConfigurator(IAppConfigurator appConfigurator)
+        {
+            if (_appConfigurator != null)
+            {
+                throw new InvalidOperationException("必须在调用 GetAppConfigurator 方法之前设置配置，请将设置配置的代码放在程序运行最前");
+            }
+
+            _appConfigurator = appConfigurator;
+        }
+
+        /// <summary>
+        /// 获取应用配置
+        /// </summary>
+        /// <returns></returns>
         public static IAppConfigurator GetAppConfigurator()
         {
             if (_appConfigurator is null)
